Abstracted lifecycle
moved lifecycle code into engine.nim. Changed from event polling to
waiting for event to reduce cpu usage and align with the ideals of a
text-editor rather than a graphics application. There seems to be an
issue with getting touchpad events this way. Will be debugging with
mouse later today.